requirements.txt
flask
flask_cors
pymssql
pycrypto
flask-cors